remove bold formatting from headings in README.md
Best practice is to allow the rendering tool (browser/whatever) to
interpret and render the heading format style. It's OK to put things
in ** for which the emphasis format holds meaning, but the intent of
Markdown is to separate content (what you write) from representation
as much as possible. If you want bold header text, then the place
to express that is in the style files (usually CSS).